requirements.txt
anthropic>=0.39.0
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
pyyaml>=6.0.1
fpdf2>=2.8.0
schedule>=1.2.0
pandas>=2.2.0
jinja2>=3.1.0
streamlit>=1.40.0
openai>=1.50.0
google-ads>=25.0.0
requests>=2.31.0
